dom/tests/mochitest/mouselock/test_constantxy.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=92264
-->
<head>
  <title>Test for Bug 633602</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="runTests();">
   <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
   <p id="display"></p>
   <div id="content" style="display: none">
   </div>
	
   <canvas id="canvas" width="150" height="150"></canvas>

   <pre id="test">
      <script>
         SimpleTest.waitForExplicitFinish();
         var canvas = document.getElementById("canvas");
         var pointer = navigator.pointer;
         var firstMove = true;
         var clientX1, clientY1, clientX2, clientY2;
         var screenX1, screenY1, screenX2, screenY2;
         SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

         canvas.addEventListener('mousemove', function(e) {
            if(firstMove) {
               clientX1 = e.clientX;
               clientY1 = e.clientY;
               screenX1 = e.screenX;
               screenY1 = e.screenY;
               firstMove = false;
            } else {
               clientX2 = e.clientX;
               clientY2 = e.clientY;
               screenX2 = e.screenX;
               screenY2 = e.screenY;
               document.mozCancelFullScreen();

               is(clientX1, clientX2, "clientX must remain constant");
               is(clientY1, clientY2, "clientY must remain constant");
               is(screenX1, screenX2, "screenX must remain constant");
               is(screenY1, screenY2, "screenY must remain constant");
               SimpleTest.finish();
            }
         }, false);
         
         // When fullscreen is enabled on canvas - lock the mouse 
         document.addEventListener('mozfullscreenchange', function() {
            if(document.mozFullScreen) {
               if(document.mozFullScreenElement == canvas) {
                  pointer.lock(canvas);
               }
            }
         }, false);
         
         function runTests() {
            canvas.mozRequestFullScreen();
            synthesizeMouse(canvas, 1000, 1000, {type: "mousemove"});
         }
      </script>
   </pre>
</body>
</html>
